Fix headline whitespace bug
The margin that spaces out the map and related components from the
important dates component I mistakenly put inside a media query. Move it
outside to effect all screen sizes.

Also, add a little margin to the right edge of important dates list, so
that as the content wraps the text never ends up flush with the right
edge line.