Move important dates to the sidebar
De-emphasize the important dates component by moving it to the right
sidebar. This is to put the main focus on the search and map
functionality, which is the major focus of the application.

Also remove the red background color to further de-emphasise the unit.
Red really draws focus, so remove it to give it the proper emphasis.