Fix horizontal scroll bars
The map box overlay element is causing some horizontal scrollbars by
being larger than the screen area.

This is because it is set to `width: 100%` *and* has `position:
absolute`. Absolutely positioned elements are absolute relative to their
first ancestor with `position: relative` (double relative, confusing,
yes), or `body`. Since the wrapper element for the map isn't positioned
relative, the overlay and the map itself calculate their width
differently. Adding `position: relative` to the wrapper fixes that up.

In addition, use `rgba` instead of `opacity` here for the transparency
effect on the overlay. This is because opacity effects the entire
element and all it's contents - including the text - where as the rgba
rule can be applied only to the background, making the definition
between background and text more clear.